SideBar, SideBarLeft, SideBarBottom components, moved logic into sidebar, separated styles into other two
